Use normal facility page for single-facility user dashboard
