+++ components/navigation/Sidebar.vue
@@ -96,15 +96,23 @@ export default defineComponent({
   },
 
   setup() {
-    const { hasPermission, hasMultipleFacilities } = usePermissions();
+    const { hasPermission, availableFacilities, firstFacility, hasMultipleFacilities } = usePermissions();
 
     const pages = computed(() => {
       return [
+        // Dashboard for multi-facility users
         {
           title: "Dashboard",
           url: "/",
           svg: "M3 12l2-2m0 0l7-7 7 7M5 10v10a1 1 0 001 1h3m10-11l2 2m-2-2v10a1 1 0 01-1 1h-3m-6 0a1 1 0 001-1v-4a1 1 0 011-1h2a1 1 0 011 1v4a1 1 0 001 1m-6 0h6",
-          visible: true,
+          visible: hasMultipleFacilities.value,
+        },
+        // Dashboard for single-facility users
+        {
+          title: "Dashboard",
+          url: firstFacility.value ? `/facilities/${firstFacility.value}` : "/",
+          svg: "M3 12l2-2m0 0l7-7 7 7M5 10v10a1 1 0 001 1h3m10-11l2 2m-2-2v10a1 1 0 01-1 1h-3m-6 0a1 1 0 001-1v-4a1 1 0 011-1h2a1 1 0 011 1v4a1 1 0 001 1m-6 0h6",
+          visible: !hasMultipleFacilities.value && availableFacilities.value?.length === 1,
         },
         {
           title: "Records",
