+++ pages/index.vue
@@ -7,16 +7,13 @@
 
     <AdminDashboard v-if="isAdmin" class="mb-6" />
 
-    <div v-if="hasMultipleFacilities">
-      <TextH2 class="mb-4">My Facilities</TextH2>
-      <FacilitiesTable @select="$router.push({ path: `/facilities/${$event}` })" />
-    </div>
-    <FacilitiesOverview v-else-if="availableFacilities.length > 0" :code="availableFacilities[0]" />
+    <TextH2 class="mb-4">My Facilities</TextH2>
+    <FacilitiesTable @select="$router.push({ path: `/facilities/${$event}` })" />
   </div>
 </template>
 
 <script lang="ts">
-import { computed, defineComponent, onMounted, ref, useContext } from "@nuxtjs/composition-api";
+import { defineComponent, onMounted, ref, useContext, useRouter } from "@nuxtjs/composition-api";
 import { DashResponse } from "@/interfaces/dash";
 
 import fetchDash from "~/helpers/fetch/fetchDash";
@@ -25,19 +22,23 @@ import usePermissions from "~/helpers/usePermissions";
 
 export default defineComponent({
   setup() {
+    const router = useRouter();
     const { $okta } = useContext();
-    const { isAdmin, getFacilities, hasMultipleFacilities } = usePermissions();
+    const { isAdmin, availableFacilities, hasMultipleFacilities } = usePermissions();
     const { fetchDashboard } = fetchDash();
 
     // Data refs
 
     const dash = ref<DashResponse>();
 
-    const availableFacilities = computed(() => {
-      return getFacilities();
-    });
+    function redirectToOnlyFacility() {
+      if (availableFacilities.value?.length === 1) {
+        router.push({ path: `/facilities/${availableFacilities.value[0]}` });
+      }
+    }
 
     onMounted(async () => {
+      redirectToOnlyFacility();
       if (await $okta.isAuthenticated()) {
         dash.value = await fetchDashboard();
       }
